subs/app/requirements.txt

fastapi==0.124.0
uvicorn[standard]==0.23.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.32.0
jinja2==3.1.3
email-validator==2.3.0
python-multipart==0.0.6
alembic
